+++ pbio/ribo/pgrms/estimate_metagene_profile_bayes_factors.py
@@ -16,26 +16,24 @@ from pbio.misc.suppress_stdout_stderr import suppress_stdout_stderr
 
 logger = logging.getLogger(__name__)
 
-default_num_cpus = 1
-
-default_periodic_models = []
-default_nonperiodic_models = []
+# When called from the Rp-Bp pipeline (create-orf-profiles), default
+# options (or else specified via the configuration file) are always
+# passed as arguments.
 
 default_periodic_offset_start = -20
 default_periodic_offset_end = 0
 default_metagene_profile_length = 21
 
-default_type_field = 'type'
-default_position_field = 'position'
-default_count_field = 'count'
-
 default_iterations = 500
 default_chains = 2
 default_seed = 8675309
 
-class NullDevice():
-    def write(self, s):
-        pass
+# Not passed as arguments, unlikely to be required
+
+default_type_field = 'type'
+default_position_field = 'position'
+default_count_field = 'count'
+
 
 def estimate_marginal_likelihoods(signal, periodic_models, nonperiodic_models, iterations, chains, seed):
         
@@ -58,13 +56,14 @@ def estimate_marginal_likelihoods(signal, periodic_models, nonperiodic_models, i
     # get the likelihood for each of the models 
     bft_periodic = [
         pm.sampling(data=data, iter=iterations, chains=chains, n_jobs=1, seed=seed, refresh=-1)
-            for pm in periodic_models]
+        for pm in periodic_models]
 
     bft_nonperiodic = [
         nm.sampling(data=data, iter=iterations, chains=chains, n_jobs=1, seed=seed, refresh=-1)
-            for nm in nonperiodic_models]
+        for nm in nonperiodic_models]
         
     return (bft_periodic, bft_nonperiodic)
+
 
 def estimate_profile_bayes_factors(profile, args):
     length = profile['length'].iloc[0]
@@ -82,14 +81,14 @@ def estimate_profile_bayes_factors(profile, args):
     start_counts = start_profile_df[args.count_field].values
     
     # find the positions of the offsets of interest within the arrays
-    begin_start_pos = np.where(start_positions==args.periodic_offset_start)[0]
+    begin_start_pos = np.where(start_positions == args.periodic_offset_start)[0]
     if len(begin_start_pos) == 0:
         msg = "Did not find any start offsets for length: {}".format(length)
         logging.warning(msg)
         return None
     begin_index = begin_start_pos[0]
 
-    stop_start_pos = np.where(start_positions==args.periodic_offset_end)[0]
+    stop_start_pos = np.where(start_positions == args.periodic_offset_end)[0]
     if len(stop_start_pos) == 0:
         msg = "Did not find any stop offsets for length: {}".format(length)
         logging.warning(msg)
@@ -107,14 +106,16 @@ def estimate_profile_bayes_factors(profile, args):
 
         # pull out the signal for this offset
         signal = start_counts[i:i+args.metagene_profile_length]
-        (bft_periodic, bft_nonperiodic) = estimate_marginal_likelihoods(signal, 
-            periodic_models, nonperiodic_models, 
-            iterations=args.iterations,chains=args.chains,seed=args.seed)
+        (bft_periodic, bft_nonperiodic) = estimate_marginal_likelihoods(signal,
+                                                                        periodic_models,
+                                                                        nonperiodic_models,
+                                                                        iterations=args.iterations,
+                                                                        chains=args.chains,
+                                                                        seed=args.seed)
     
         # extract the parameters of interest
         m_periodic_ex = [m.extract(pars=['lp__']) for m in bft_periodic]
         m_nonperiodic_ex = [m.extract(pars=['lp__']) for m in bft_nonperiodic]
-        
 
         # now, choose the best model of each class,  based on mean likelihood
         m_periodic_means = [np.mean(m_ex['lp__']) for m_ex in m_periodic_ex]
@@ -150,49 +151,55 @@ def estimate_profile_bayes_factors(profile, args):
 
     return ret
 
+
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        description="This script estimates the Bayes factors for all metagene profiles in the "
-        "given file. The script accepts as input multiple \"periodic\" and \"nonperiodic\" "
-        "models. It uses the models of each type with the best mean to estimate the Bayes "
-        "factor distributions.\n\nIt contains some hard-coded field names.")
+                                     description="""This script estimates the Bayes factors 
+        for all metagene profiles in the given file. The script accepts as input multiple 
+        'periodic' and 'nonperiodic' models. It uses the models of each type with the 
+        best mean to estimate the Bayes factor distributions. It contains some 
+        hard-coded field names.""")
         
     parser.add_argument('metagene_profiles', help="The (csv) file containing the metagene profiles")
+
     parser.add_argument('out', help="The output (csv.gz) file")
 
-    parser.add_argument('--periodic-models', help="A list of pickled StanModel files which contain "
-        "models that somehow represent periodic metagene profiles", nargs="+", 
-        default=default_periodic_models)
-    parser.add_argument('--nonperiodic-models', help="A list of pickled StanModel files which contain "
-        "models that somehow represent nonperiodic metagene profiles", nargs="+", 
-        default=default_nonperiodic_models)
+    parser.add_argument('--periodic-models', help="""A list of pickled StanModel files which contain
+        models that somehow represent periodic metagene profiles.""", nargs="+", default=[])
 
+    parser.add_argument('--nonperiodic-models', help=""""A list of pickled StanModel files which contain
+        models that somehow represent nonperiodic metagene profiles.""", nargs="+", default=[])
 
-    parser.add_argument('--periodic-offset-start', help="The position, relative to the translation "
-        "initiation site, to begin calculating periodicity Bayes factors (inclusive)", type=int,
-        default=default_periodic_offset_start)
-    parser.add_argument('--periodic-offset-end', help="The position, relative to the translation "
-        "initiation site, to stop calculating periodicity Bayes factors (inclusive)", type=int,
-        default=default_periodic_offset_end)
-    parser.add_argument('--metagene-profile-length', help="The length of the profile to use in the "
-        "models. metagene_profile_length + periodic_offset_end must be consistent with the length "
-        "of the extracted metagene profile. The length must be divisible by three.", type=int,
-        default=default_metagene_profile_length)
+    parser.add_argument('--periodic-offset-start', help="""The position, relative to the translation
+        initiation site, to begin calculating periodicity Bayes factors (inclusive).""", type=int,
+                        default=default_periodic_offset_start)
 
+    parser.add_argument('--periodic-offset-end', help="""The position, relative to the translation
+        initiation site, to stop calculating periodicity Bayes factors (inclusive)""", type=int,
+                        default=default_periodic_offset_end)
+
+    parser.add_argument('--metagene-profile-length', help="""The length of the profile to use in the
+        models. metagene_profile_length + periodic_offset_end must be consistent with the length
+        of the extracted metagene profile. The length must be divisible by three.""", type=int,
+                        default=default_metagene_profile_length)
 
     parser.add_argument('-s', '--seed', help="The random seeds to use for inference",
-        type=int, default=default_seed)
+                        type=int, default=default_seed)
+
     parser.add_argument('-c', '--chains', help="The number of MCMC chains to use", type=int,
-        default=default_chains)
-    parser.add_argument('-i', '--iterations', help="The number of MCMC iterations to use for "
-        "each chain", type=int, default=default_iterations)
+                        default=default_chains)
+
+    parser.add_argument('-i', '--iterations', help="""The number of MCMC iterations to use for
+        each chain.""", type=int, default=default_iterations)
     
-    parser.add_argument('-p', '--num-cpus', help="The number of CPUs to use. Each read "
-        "length will be processed in its own thread (so that is the maximum number of CPUs "
-        "that is useful).", type=int, default=default_num_cpus)
+    parser.add_argument('-p', '--num-cpus', help="""The number of CPUs to use. Each read
+        length will be processed in its own thread (so this is the maximum number of CPUs
+        that is useful).""", type=int, default=1)
 
     parser.add_argument('--type-field', default=default_type_field)
+
     parser.add_argument('--count-field', default=default_count_field)
+
     parser.add_argument('--position-field', default=default_position_field)
 
     logging_utils.add_logging_options(parser)
@@ -228,6 +235,6 @@ def main():
 
     pandas_utils.write_df(all_profile_estimates_df, args.out, index=False)
 
+
 if __name__ == '__main__':
     main()
-