+++ pbio/ribo/pgrms/extract_metagene_profiles.py
@@ -1,5 +1,13 @@
 #! /usr/bin/env python3
 
+"""Extract metagene profiles.
+
+Functions:
+    get_interval_df
+    get_length_strand_profiles
+    get_metagene_profile_df
+"""
+
 import argparse
 import collections
 import numpy as np
@@ -12,11 +20,12 @@ import pbio.misc.pandas_utils as pandas_utils
 
 import logging
 import pbio.misc.logging_utils as logging_utils
+
 logger = logging.getLogger(__name__)
 
-
-default_num_cpus = 1
-default_lengths = []
+# When called from the Rp-Bp pipeline (create-orf-profiles), default
+# options (or else specified via the configuration file) are always
+# passed as arguments.
 
 default_start_upstream = 300
 default_start_downstream = 300
@@ -25,6 +34,7 @@ default_end_downstream = 300
 
 
 def get_interval_df(start, end, seqname, strand):
+
     interval_df = pd.DataFrame()
     interval_df['start'] = start
     interval_df['end'] = end
@@ -32,9 +42,12 @@ def get_interval_df(start, end, seqname, strand):
     interval_df['strand'] = strand
     interval_df['id'] = "."
     interval_df['score'] = 0
+
     return interval_df
 
+
 def get_length_strand_profiles(matches, profile_length):
+
     init = lambda: np.zeros(profile_length, int)
     length_strand_profiles = collections.defaultdict(init)
     
@@ -51,7 +64,9 @@ def get_length_strand_profiles(matches, profile_length):
         
     return length_strand_profiles
 
+
 def get_metagene_profile_df(length, type_label, length_strand_profiles, upstream, downstream):
+
     reverse_metagene_profile = length_strand_profiles[(length, '-')]
     forward_metagene_profile = length_strand_profiles[(length, '+')]
     
@@ -69,39 +84,45 @@ def get_metagene_profile_df(length, type_label, length_strand_profiles, upstream
     metagene_profile_df['length'] = length
     
     return metagene_profile_df
-     
+
+
 def main():
 
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        description="This script extracts the metagene profile from reads in a BAM "
-        "file, possibly filtering by length. It attempts to vectorize as many of the "
-        "counting operations as possible.")
+                                     description="""This script extracts the metagene profile 
+        from reads in a BAM file, possibly filtering by length.""")
+
     parser.add_argument('bam', help="The bam file")
+
     parser.add_argument('orfs', help="The annotated transcripts (bed) file")
+
     parser.add_argument('out', help="The (output) csv.gz counts file")
 
     parser.add_argument('-p', '--num-cpus', help="The number of processors to use",
-        type=int, default=default_num_cpus)
+                        type=int, default=1)
 
-    parser.add_argument('--is-sam', help="If this flag is present, the alignment file will "
-        "be parsed as SAM rather than BAM", action='store_true')
+    parser.add_argument('--is-sam', help="""If this flag is present, the alignment file will
+        be parsed as SAM rather than BAM""", action='store_true')
 
-    parser.add_argument('--lengths', help="If specified, then metagene profiles will be "
-        "created for reads of each length. Otherwise, profiles will be created for each "
-        "read length present in the bam file.", type=int, nargs='*', default=default_lengths)
+    parser.add_argument('--lengths', help="""If specified, then metagene profiles will be
+        created for reads of each length. Otherwise, profiles will be created for each
+        read length present in the bam file.""", type=int, nargs='*', default=[])
 
-    parser.add_argument('--start-upstream', type=int, default=default_start_upstream, 
-        help="The number of bases upstream of the translation initiation site to begin "
-        "constructing the metagene profile.")
+    parser.add_argument('--start-upstream', type=int, default=default_start_upstream,
+                        help="""The number of bases upstream of the translation initiation 
+        site to begin constructing the metagene profile.""")
+
     parser.add_argument('--start-downstream', type=int, default=default_start_downstream,
-        help="The number of bases downstream of the translation initiation site to end "
-        "the metagene profile.")
+                        help="""The number of bases downstream of the translation initiation 
+        site to end the metagene profile.""")
+
     parser.add_argument('--end-upstream', type=int, default=default_end_upstream,
-        help="The number of bases upstream of the translation termination site to begin "
-        "constructing the metagene profile.")
+                        help="""The number of bases upstream of the translation termination 
+        site to begin constructing the metagene profile.""")
+
     parser.add_argument('--end-downstream', type=int, default=default_end_downstream,
-        help="The number of bases downstream of the translation termination site to end "
-        "the metagene profile.")
+                        help="""The number of bases downstream of the translation termination 
+        site to end the metagene profile.""")
     
     logging_utils.add_logging_options(parser)
     args = parser.parse_args()
@@ -111,8 +132,9 @@ def main():
     logger.info(msg)
 
     # first, get the 5' ends of the reads
-    alignment_df = bam_utils.get_five_prime_ends(args.bam, progress_bar=True, 
-        count=True, logger=logger)
+    alignment_df = bam_utils.get_five_prime_ends(args.bam,
+                                                 progress_bar=True,
+                                                 count=True, logger=logger)
 
     msg = "Reading annotations"
     logger.info(msg)
@@ -198,7 +220,6 @@ def main():
     msg = "Extracting metagene profiles"
     logger.info(msg)
 
-    
     if len(args.lengths) == 0:
         args.lengths = list(alignment_df['length'].unique())
 
@@ -221,11 +242,11 @@ def main():
         all_metagene_profile_dfs.append(initiation_profile_df)
         
         # and around termination sites
-        termination_profile_df = get_metagene_profile_df(length, 
-                                                        'end',
-                                                        termination_length_strand_profiles, 
-                                                        args.end_upstream, 
-                                                        args.end_downstream)
+        termination_profile_df = get_metagene_profile_df(length,
+                                                         'end',
+                                                         termination_length_strand_profiles,
+                                                         args.end_upstream,
+                                                         args.end_downstream)
         
         all_metagene_profile_dfs.append(termination_profile_df)
         
@@ -238,6 +259,7 @@ def main():
     msg = "Writing metagene profiles to disk"
     logger.info(msg)
     pandas_utils.write_df(all_metagene_profile_dfs, args.out, index=False)
+
 
 if __name__ == '__main__':
     main()